avahi: Drop package
The daemon locks up when starting up in avahi_log_info() and
probably the other logging functions, too.

Since avahi is not really used a lot in the distribution,
has been in testing for four years and has virtually no users
I am going to drop it instead of wasting time on fixing this.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>