update nagios_nrpe to 3.2.1
Update NRPE to 3.2.1, which fixes various security vulnerabilities
and bugs (changelog is available at:
https://github.com/NagiosEnterprises/nrpe/blob/master/CHANGELOG.md).

Due to changes in ./configure, some options such as initsystem (sysV)
need to be specified now. Also, simple "make" does not exist anymore
and must be replaced by "make all".

Rootfile information is now included which was missing in the
first version of this patch.

This fixes #11700 and partly addresses #11551.

Signed-off-by: Peter Müller <peter.mueller@link38.eu>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>