make.sh: Build in ramdisk
This is an experimental change that I want to trial to speed up
the nightly builds. The build environment will be mounted in a
ramdisk and the build will be performed in there.

This will hopefully reduce IO on the (slow) replicated disks.

If there is no significant performance gain from this, this
commit will be reverted.

To enable this, USE_RAMDISK must be set to 1 in .config.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>